product/requirements.txt
fastapi
uvicorn
pymongo
python-jose
pycrptodome
openpyxl
requests
pytho-dotenv